+++ package.json
@@ -28,6 +28,7 @@
     "@mantine/prism": "^6.0.10",
     "@mantine/spotlight": "^6.0.10",
     "@mantine/tiptap": "^6.0.10",
+    "@mozilla/readability": "^0.4.4",
     "@next-auth/prisma-adapter": "^1.0.5",
     "@prisma/client": "^4.11.0",
     "@supabase/supabase-js": "^2.22.0",
@@ -48,6 +49,7 @@
     "daisyui": "^2.51.6",
     "dayjs": "^1.11.7",
     "embla-carousel-react": "^7.1.0",
+    "eventsource-parser": "^1.0.0",
     "i18next": "^22.5.0",
     "next": "13.3.4",
     "next-auth": "^4.21.0",
