+++ src/pages/api/chat.ts
@@ -4,7 +4,7 @@ import { OpenAIError, OpenAIStream } from '@/utils/server'
 import { ChatBody, Message } from '@/types/chat'
 
 // @ts-expect-error
-import wasm from '../../node_modules/@dqbd/tiktoken/lite/tiktoken_bg.wasm?module'
+import wasm from '../../../node_modules/@dqbd/tiktoken/lite/tiktoken_bg.wasm?module'
 
 import tiktokenModel from '@dqbd/tiktoken/encoders/cl100k_base.json'
 import { Tiktoken, init } from '@dqbd/tiktoken/lite/init'
@@ -42,14 +42,27 @@ const handler = async (req: Request): Promise<Response> => {
 
     for (let i = messages.length - 1; i >= 0; i--) {
       const message = messages[i]
-      const tokens = encoding.encode(message.content)
+      if (message) {
+        const tokens = encoding.encode(message.content)
 
-      if (tokenCount + tokens.length + 1000 > model.tokenLimit) {
-        break
+        if (tokenCount + tokens.length + 1000 > model.tokenLimit) {
+          break
+        }
+        tokenCount += tokens.length
+        messagesToSend = [message, ...messagesToSend]
       }
-      tokenCount += tokens.length
-      messagesToSend = [message, ...messagesToSend]
     }
+
+    // for (let i = messages.length - 1; i >= 0; i--) {
+    //   const message = messages[i]
+    //   const tokens = encoding.encode(message.content)
+
+    //   if (tokenCount + tokens.length + 1000 > model.tokenLimit) {
+    //     break
+    //   }
+    //   tokenCount += tokens.length
+    //   messagesToSend = [message, ...messagesToSend]
+    // }
 
     encoding.free()
